+++ trlc/ast.py
@@ -1529,6 +1529,10 @@ class Builtin_Function(Entity):
 
     :attribute arity: number of parameters
     :type: int
+
+    :attribute deprecated: if this functions is deprecated and should no \
+    longer be used
+    :type: bool
     """
     LOCATION = Location(file_name = "<builtin>")
 
@@ -1537,6 +1541,7 @@ class Builtin_Function(Entity):
         assert isinstance(arity, int)
         assert arity >= 0
         self.arity = arity
+        self.deprecated = ":" in name
 
     def dump(self, indent=0):  # pragma: no cover
         self.write_indent(indent, self.__class__.__name__ + " " + self.name)